Se cambia tipo de dato

+++ backend/app/src/main/java/mims/app/entity/JoyaEntity.java
@@ -24,7 +24,7 @@ public class JoyaEntity {
     private int id_tipo_joya;
 
     @Column(name = "cost")
-    private int cost;
+    private float cost;
 
     @Column(name = "deleted")
     private boolean deleted;
